Remove AutoMapper (#141)
* Remove AutoMapper

* Removed AutoMapper. Part 2.

+++ src/Steam.Models/PublishedFileDetailsModel.cs
@@ -17,7 +17,7 @@ namespace Steam.Models
 
         public string FileName { get; set; }
 
-        public uint FileSize { get; set; }
+        public ulong FileSize { get; set; }
 
         public Uri FileUrl { get; set; }
